Working prototype of NPEC parallelized script with test case
