+++ requirements.txt
@@ -8,6 +8,7 @@ matplotlib!=3.3.1,!=3.3.0
 numpy
 pandas
 pymdptoolbox
+ray>=1.0.0
 seaborn
 setuptools
 scikit-learn
