+++ tests/test_scripts.py
@@ -27,8 +27,8 @@ from evaluating_rewards.analysis.distances import (
     table_combined,
 )
 from evaluating_rewards.analysis.reward_figures import plot_gridworld_reward, plot_pm_reward
-from evaluating_rewards.scripts import npec_comparison, train_preferences, train_regress
-from evaluating_rewards.scripts.distances import epic, erc
+from evaluating_rewards.scripts import train_preferences, train_regress
+from evaluating_rewards.scripts.distances import epic, erc, npec
 from tests import common
 
 
@@ -64,6 +64,15 @@ EXPERIMENTS = {
     # experiment, expected_type, extra_named_configs, config_updates, extra_check
     "epic_distance": (epic.epic_distance_ex, dict, [], {}, _check_distance_return),
     "erc_distance": (erc.erc_distance_ex, dict, [], {}, _check_distance_return),
+    "npec_distance": (
+        npec.npec_distance_ex,
+        dict,
+        [],
+        {
+            "num_cpus": 2,  # CI build only has 2 cores
+        },
+        _check_distance_return,
+    ),
     "erc_distance_spearman": (
         erc.erc_distance_ex,
         dict,
@@ -94,9 +103,9 @@ EXPERIMENTS = {
     ),
     "plot_pm_reward": (plot_pm_reward.plot_pm_reward_ex, xr.DataArray, [], {}, None),
     "table_combined": (table_combined.table_combined_ex, type(None), [], {}, None),
-    "comparison": (npec_comparison.npec_comparison_ex, dict, [], {}, None),
+    "comparison": (npec.npec_distance_ex, dict, [], {}, None),
     "comparison_alternating": (
-        npec_comparison.npec_comparison_ex,
+        npec.npec_distance_ex,
         dict,
         ["alternating_maximization"],
         {"fit_kwargs": {"epoch_timesteps": 4096}},
